generateapi/requirements.txt
Flask
psycopg2-binary
sentence-transformers
transformers
numpy